Move taxon routes to own file
